feat: replace legacy tracetrace with rtracetrace (#2159)
This puts through a fairly sizeable refactoring to use rtrace.Trace
everywhere in place of trace.Trace.

Key changes:

* Use rtrace.Trace everywhere in place of trace.Trace; move the JSON
  reader/writer from pkg/trace/json/ to pkg/rtrace/json/.
* Simplify ComputedRegister (pkg/ir/assignment/computed_register.go)
  and trim related assignment utilities (pseudo_inverse, util).
* Remove padding from register.Register, since rtrace.Trace does not
  support padding. Delete pkg/ir/padding.go and
  pkg/ir/trace_alignment.go; introduce pkg/ir/builder/alignment.go
  as the replacement.
* Delete the field-agnostic schema layer (pkg/schema/agnostic/),
  including assignment, equation, filler, polynomial, and util types.
* Slim pkg/schema/register/: remove the register allocator, constant
  register, and vector types.
* Remove pkg/util/collection/array/ (dense/sparse codecs, builders,
  pool/static/small/vec arrays) and pkg/util/collection/pool/
  (object-pool utilities). Expand pkg/util/collection/narray/ to
  absorb the needed functionality.
* Gut the legacy pkg/trace/ package (array_trace, ref, trace, util).
* Update all downstream consumers: CLI commands (check, debug,
  inspect, trace, zkc/*), AIR gadgets (bitwidth, normalisation),
  MIR concretization, IR terms, schema constraints, and ZkC
  constraints to work against the new rtrace.Trace interface.

* feat: improve performance of alignment / padding

This improves the memory performance of trace alignment / padding by
avoiding the need to clone the memory twice over.  Specifically, this is
done by merging Clone into Pad(), such that the clone happens just once.

Furthermore, this tidies up some aspects of the package structure by
merging previously duplicated packages.

---------

+++ pkg/schema/constraint/lookup/constraint.go
@@ -194,18 +194,18 @@ func (p Constraint[F]) checkSourceSet(src SetId, mod trace.Module[F], sets []Set
 	if src.HasSelector() {
 		var selector = src.Selector().Unwrap()
 		//
-		for row := range int(mod.Height()) {
+		for row := range mod.Height() {
 			if !mod.Column(selector).Get(row).IsZero() {
 				if !contains(row, src, mod, sets, buffer) {
-					return &Failure[F]{p.Handle, src, uint(row)}
+					return &Failure[F]{p.Handle, src, row}
 				}
 			}
 		}
 	} else {
 		// Optimised path when no selector
-		for row := range int(mod.Height()) {
+		for row := range mod.Height() {
 			if !contains(row, src, mod, sets, buffer) {
-				return &Failure[F]{p.Handle, src, uint(row)}
+				return &Failure[F]{p.Handle, src, row}
 			}
 		}
 	}
@@ -216,7 +216,7 @@ func (p Constraint[F]) checkSourceSet(src SetId, mod trace.Module[F], sets []Set
 // check whether the given source row is contained within any of the given sets.
 // A temporary buffer of sufficient width is provided to avoid memory
 // allocation.
-func contains[F field.Element[F]](row int, src SetId, mod trace.Module[F], sets []Set[[]F], buffer []F) bool {
+func contains[F field.Element[F]](row uint, src SetId, mod trace.Module[F], sets []Set[[]F], buffer []F) bool {
 	// Read registers into buffer
 	for i := range src.Width() {
 		var rid = src.Ith(i).Unwrap()
